Moved cache dto to common-data
