+++ nolisservices/common-data/src/main/java/com/nolis/commondata/dto/CacheDTO.java
@@ -1,0 +1,8 @@
+package com.nolis.commondata.dto;
+
+public record CacheDTO(
+        String name,
+        String ttl,
+        String unit
+) {
+}
